requirements.txt
streamlit>=1.38.0
openai>=1.40.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0 # For reading PDFs locally
python-docx>=1.1.0 # For reading DOCX locally
